# docs/openapi.yaml
openapi: 3.1.0
info:
  title: LMS V1 API
  version: 1.0.0
  description: Canonical single-owner LMS contract. NestJS is the authorization authority.
  license:
    name: UNLICENSED
    identifier: LicenseRef-Proprietary
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
...
security: []
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      operationId: registerStudent
      summary: Register a student account
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterRequest" }
      responses:
        "201":
          description: Student registered
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "409": { $ref: "#/components/responses/Conflict" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: Authenticate and register/resolve a device
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: Authenticated; rotating refresh cookie is set
          headers:
            Set-Cookie:
              schema: { type: string }
              description: HttpOnly refresh cookie scoped to the refresh endpoint
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResponse" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /auth/refresh:
    post:
      tags: [Auth]
      operationId: refreshSession
      summary: Rotate refresh token and issue an access token
      security: [{ refreshCookie: [] }]
      responses:
        "200":
          description: Token rotated
          headers:
            Set-Cookie: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /auth/logout:
    post:
      tags: [Auth]
      operationId: logout
      summary: Revoke the current session
      security: [{ bearerAuth: [] }, { refreshCookie: [] }]
      responses:
        "204": { description: Logged out }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/logout-all:
    post:
      tags: [Auth]
      operationId: logoutAll
      summary: Revoke all subject sessions and playback
      security: [{ bearerAuth: [] }]
      responses:
        "204": { description: All sessions revoked }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/password/forgot:
    post:
      tags: [Auth]
      operationId: forgotPassword
      summary: Request password reset without account enumeration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties: { email: { type: string, format: email } }
      responses:
        "202": { description: Request accepted whether or not the account exists }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /auth/password/reset:
    post:
      tags: [Auth]
      operationId: resetPassword
      summary: Consume a one-time reset token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, newPassword]
              properties:
                token: { type: string, minLength: 32 }
                newPassword: { type: string, minLength: 12, maxLength: 128 }
      responses:
        "204": { description: Password changed and sessions revoked }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /auth/password/change:
    post:
      tags: [Auth]
      operationId: changePassword
      summary: Change the current subject password
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword: { type: string }
                newPassword: { type: string, minLength: 12, maxLength: 128 }
      responses:
        "204": { description: Password changed; other sessions revoked }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /me:
    get:
      tags: [Account]
      operationId: getMe
      summary: Get the current subject profile
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    patch:
      tags: [Account]
      operationId: updateMe
      summary: Update the current subject profile
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              minProperties: 1
              properties:
                firstName: { type: string, minLength: 1, maxLength: 100 }
                lastName: { type: string, minLength: 1, maxLength: 100 }
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /me/devices:
    get:
      tags: [Account]
      operationId: listMyDevices
      summary: List the current subject devices
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Subject devices
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Device" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /me/devices/{deviceId}:
    delete:
      tags: [Account]
      operationId: revokeMyDevice
      summary: Revoke one current-subject device
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/DeviceId"
      responses:
        "204": { description: Device and its sessions revoked }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: lms_refresh
  parameters:
...
    DeviceId: { name: deviceId, in: path, required: true, schema: { $ref: "#/components/schemas/Uuid" } }
...
  responses:
    ValidationFailed:
      description: Request validation failed
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Unauthorized:
      description: Authentication failed
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Forbidden:
      description: Authorization or account policy denied access
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    NotFound:
      description: Resource not found or concealed by authorization
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Conflict:
      description: State or uniqueness conflict
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After: { schema: { type: integer } }
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
  schemas:
    Uuid: { type: string, format: uuid }
    Timestamp: { type: string, format: date-time }
    Error:
      type: object
      additionalProperties: false
      required: [statusCode, code, message, requestId]
      properties:
        statusCode: { type: integer, minimum: 400, maximum: 599 }
        code: { type: string, pattern: "^[A-Z][A-Z0-9_]+$" }
        message: { type: string }
        details: { type: object, additionalProperties: true }
        requestId: { $ref: "#/components/schemas/Uuid" }
...
    RegisterRequest:
      type: object
      additionalProperties: false
      required: [email, password, firstName, lastName]
      properties:
        email: { type: string, format: email, maxLength: 320 }
        password: { type: string, minLength: 12, maxLength: 128 }
        firstName: { type: string, minLength: 1, maxLength: 100 }
        lastName: { type: string, minLength: 1, maxLength: 100 }
    LoginRequest:
      type: object
      additionalProperties: false
      required: [email, password, device]
      properties:
        email: { type: string, format: email, maxLength: 320 }
        password: { type: string, maxLength: 128 }
        device: { $ref: "#/components/schemas/DeviceInput" }
    DeviceInput:
      type: object
      additionalProperties: false
      required: [clientDeviceId, name]
      properties:
        clientDeviceId: { $ref: "#/components/schemas/Uuid" }
        name: { type: string, minLength: 1, maxLength: 120 }
    TokenResponse:
      type: object
      additionalProperties: false
      required: [accessToken, accessTokenExpiresIn]
      properties:
        accessToken: { type: string }
        accessTokenExpiresIn: { type: integer, minimum: 1, example: 900 }
    AuthResponse:
      allOf:
        - $ref: "#/components/schemas/TokenResponse"
        - type: object
          required: [user, device]
          properties:
            user: { $ref: "#/components/schemas/User" }
            device: { $ref: "#/components/schemas/Device" }
    User:
      type: object
      additionalProperties: false
      required: [id, email, firstName, lastName, status, roles, createdAt, updatedAt]
      properties:
        id: { $ref: "#/components/schemas/Uuid" }
        email: { type: string, format: email }
        firstName: { type: string }
        lastName: { type: string }
        status: { type: string, enum: [PENDING_VERIFICATION, ACTIVE, SUSPENDED, DISABLED] }
        roles: { type: array, items: { type: string, enum: [OWNER, STUDENT] } }
        createdAt: { $ref: "#/components/schemas/Timestamp" }
        updatedAt: { $ref: "#/components/schemas/Timestamp" }
    Device:
      type: object
      additionalProperties: false
      required: [id, clientDeviceId, name, firstSeenAt, lastSeenAt, revoked]
      properties:
        id: { $ref: "#/components/schemas/Uuid" }
        clientDeviceId: { $ref: "#/components/schemas/Uuid" }
        name: { type: string }
        browser: { type: [string, "null"] }
        operatingSystem: { type: [string, "null"] }
        firstSeenAt: { $ref: "#/components/schemas/Timestamp" }
        lastSeenAt: { $ref: "#/components/schemas/Timestamp" }
        revoked: { type: boolean }
...